Vain Riser Project (#107)
* AVGPlayer

* Bug Fixed.

* Command Processing

* Bug Fixed.

* Bug Fixed

* AVG Player

* Bug Fixed

* Bug Fixed.

* Font Bug Fixed

* Fade Font Animation

* Bug Fixed

* AVG Components

* Async Launcher

* Engine Construct Bug Fixed

* Bug Fixed

* Bug Fixed

* Remove Steam

* Bug Fixed.

* AsyncLauncher Cancel

* Bug Fixed

* Bug Fixed

* More Convert

* Bug Fixed.

* Yukimi::TextWindow::AppendWaitTime

* Bug Fixed

* Vec3

* Bug Fixed

* Bug Fixed

* Bug Fixed.

* Bug fixed.

* Bug Fixed

* Update

* Design Bug Fixed.

* Modern Styles

* Update

* Bug Fixed.

* Bug Fixed

* Bug Fixe

* Bug Fixed

* Bug Fixed

* Optimized

* Metronome

* Bug Fixed

* Sound BPM MetaData

* BGM Metadata

* Bug Fixed

* Bug Fixed

* Bug Fixed

+++ Core/RenderTarget.h
@@ -16,7 +16,7 @@ namespace Snowing::Graphics
 			impl_{ std::move(impl) }
 		{}
 
-		RenderTargetInterface(Math::Vec2<int> size) :
+		RenderTargetInterface(Math::Vec2<size_t> size) :
 			impl_{ size }
 		{}
